Extract process plumbing; drop pub(crate); resolve() returns just the CLI
Three review follow-ups:

- Move the engine-agnostic child-process plumbing (output_with_timeout,
  ProcessGroupChild) out of container.rs into a new `process` module. Nothing
  there knows about docker/podman; the container runtime, compose, and the k8s
  validator all build on it. container.rs drops ~140 lines.

- Replace pub(crate) with pub in these modules — snouty is a binary, not a
  published library, so the narrower visibility bought nothing.

- ComposeCli::resolve() no longer returns a (cli, version) tuple. Resolving and
  reading the version are separate concerns: resolve() returns the CLI (running
  `version` only to confirm v2), and callers that want the banner call
  .version() on the result. `probe()` does exactly that.

426 tests pass, clippy clean, doctor verified.
